requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
Pillow>=8.3.1
scikit-learn>=0.24.2
matplotlib>=3.4.3
pandas>=1.3.3
tqdm>=4.62.3
kaggle>=1.5.12
flask==2.0.1
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1 